Remove subfolder Post/
Getting rid of the Post subfolder so now the Post versions have replaced the versions in the parent folder. This means you now have to explicitly deliver the the normal we are perturbing and the position to GenBasisTB and the base normal also has to be given always to Resolve Surfgrad.
It makes it a bit more clunky for now until we get the ability in Shader Graph to provide default inputs to a subgraph as nodes.
On the bright side it makes less room for error because now the dependency on the initial surface normal is explicit which makes for less of a for a user.