better compression for three channel compressed normal maps
added a subgraph vec3 ts normal to deriv.subgraph which does not rely on the provided tangent space normal being unit length which generally fails when compression is enabled. To make a three channel normal map use type Default in the texture inspector and disable SRGB. Then sample the texture in your shader graph using the setting Normal and Object in the dropdowns (eventhough it's not an OS nmap). The difference is particularly noticeable with very small bump scales.